+++ asyncmy/replication/packets.py
@@ -1,4 +1,5 @@
-import xstruct as struct
+import struct
+
 from asyncmy.constants.COLUMN import (
     NULL_COLUMN,
     UNSIGNED_CHAR_COLUMN,