+++ asyncmy/replication/row_events.py
@@ -1,7 +1,7 @@
 import datetime
 import decimal
+import struct
 
-import xstruct as struct
 from asyncmy.charset import charset_by_name
 from asyncmy.constants.FIELD_TYPE import (
     BIT,