+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix `binlogstream` await. (#12)
 - Remove `loop` argument. (#15)
+- Fix `unix_socket` connect. (#17)
 
 ### 0.2.0